Non blocking capture

+++ pupil_src/shared_modules/video_capture/ndsi_backend.py
@@ -9,15 +9,15 @@ See COPYING and COPYING.LESSER for license details.
 ---------------------------------------------------------------------------~(*)
 '''
 
-import ndsi
+import time
 import logging
 
-from .base_backend import InitialisationError, StreamError, Base_Source, Base_Manager
-from .fake_backend import Fake_Source
+import ndsi
 
-assert ndsi.NDS_PROTOCOL_VERSION >= '0.2.13'
+from .base_backend import Base_Source, Base_Manager
+
+assert ndsi.NDS_PROTOCOL_VERSION >= '0.2.14'
 logger = logging.getLogger(__name__)
-logger.setLevel(logging.DEBUG)
 
 
 class NDSI_Source(Base_Source):
@@ -27,73 +27,115 @@ class NDSI_Source(Base_Source):
         get_frame_timeout (float): Maximal waiting time for next frame
         sensor (ndsi.Sensor): NDSI sensor backend
     """
-    def __init__(self, g_pool, frame_size, frame_rate, network=None, source_id=None, host_name=None, sensor_name=None, **settings):
-        if not network:
-            raise InitialisationError()
-        super(NDSI_Source, self).__init__(g_pool)
+    def __init__(self, g_pool, frame_size, frame_rate, network=None,
+                 source_id=None, host_name=None, sensor_name=None):
+        super().__init__(g_pool)
         self.sensor = None
-        try:
-            # uuid given
-            if source_id:
-                self.sensor = network.sensor(source_id, callbacks=(self.on_notification,))
-            # host/sensor name combination, prob. from settings
-            elif host_name and sensor_name:
-                for sensor in network.sensors.values():
-                    if sensor['host_name'] == host_name and sensor['sensor_name'] == sensor_name:
-                        self.sensor = network.sensor(sensor['sensor_uuid'], callbacks=(self.on_notification,))
-            else:
-                for sensor_id in network.sensors:
-                    try:
-                        self.sensor = network.sensor(sensor_id, callbacks=(self.on_notification,))
-                        break
-                    except ValueError:
-                        continue
-        except ValueError:
-            raise InitialisationError()
-        if not self.sensor:
-            raise InitialisationError()
-        if not self.sensor.supports_data_subscription:
-            self.cleanup()
-            raise InitialisationError('Source does not support data subscription.')
-
-        logger.debug('NDSI Source Sensor: {}'.format(self.sensor))
-        self.control_id_ui_mapping = {}
-        self.get_frame_timeout = 1000
-        self._frame_size = None
-        self.frame_size = frame_size
-        self.frame_rate = frame_rate
+        self._source_id = source_id
+        self._sensor_name = sensor_name
+        self._host_name = host_name
+        self._frame_size = frame_size
+        self._frame_rate = frame_rate
         self.has_ui = False
+        self.control_id_ui_mapping = {}
+        self.get_frame_timeout = 100  # ms
+        self.ghost_mode_timeout = 10  # sec
         self._initial_refresh = True
+        self.last_update = self.g_pool.get_timestamp()
+
+        if not network:
+            logger.debug('No network reference provided. Capture is started '
+                         + 'in ghost mode. No images will be supplied.')
+            return
+
+        self.recover(network)
+
+        if not self.sensor or not self.sensor.supports_data_subscription:
+            logger.error('Init failed. Capture is started in ghost mode. '
+                         + 'No images will be supplied.')
+            self.cleanup()
+
+        logger.debug('NDSI Source Sensor: %s' % self.sensor)
+
+    def recover(self, network):
+        logger.debug('Trying to recover with %s, %s, %s' % (self._source_id,
+                                                            self._sensor_name,
+                                                            self._host_name))
+        if self._source_id:
+            try:
+                # uuid given
+                self.sensor = network.sensor(self._source_id,
+                                             callbacks=(self.on_notification,))
+            except ValueError:
+                pass
+
+        if self.online:
+            self._sensor_name = self.sensor.name
+            self._host_name = self.sensor.host_name
+            return
+        if self._host_name and self._sensor_name:
+            for sensor in network.sensors.values():
+                if (sensor['host_name'] == self._host_name
+                        and sensor['sensor_name'] == self._sensor_name):
+                    self.sensor = network.sensor(
+                        sensor['sensor_uuid'],
+                        callbacks=(self.on_notification,))
+                    if self.online:
+                        self._sensor_name = self.sensor.name
+                        self._host_name = self.sensor.host_name
+                        break
+        else:
+            for s_id in network.sensors:
+                self.sensor = network.sensor(s_id,
+                                             callbacks=(self.on_notification,))
+                if self.online:
+                    self._sensor_name = self.sensor.name
+                    self._host_name = self.sensor.host_name
+                    break
 
     @property
     def name(self):
-        return '{} @ {}'.format(self.sensor.name, self.sensor.host_name)
+        return '{} @ {}'.format(self._sensor_name, self._host_name)
+
+    @property
+    def online(self):
+        return bool(self.sensor)
 
     def poll_notifications(self):
         while self.sensor.has_notifications:
             self.sensor.handle_notification()
 
-    def _get_frame(self):
-        self.poll_notifications()
-        return self.sensor.get_newest_data_frame(timeout=self.get_frame_timeout)
-
-    def get_frame(self):
-        '''Mirrors uvc.Capture.get_frame_robust()'''
-        attempts = 3
-        for a in range(attempts):
+    def recent_events(self, events):
+        if self.online:
+            self.poll_notifications()
             try:
-                frame = self._get_frame()
-                self.frame_size = (frame.width, frame.height)
-            except Exception as e:
-                if a:
-                    logger.info('Could not get Frame: "{}". Attempt: {}/{} '.format(str(e), a+1, attempts))
-                else:
-                    logger.debug('Could not get Frame of first try: "{}". Attempt: {}/{}'.format(str(e), a+1, attempts))
-            else:
-                return frame
-        raise StreamError("Could not grab frame after 3 attempts. Giving up.")
+                frame = self.sensor.get_newest_data_frame(timeout=self.get_frame_timeout)
+            except ndsi.StreamError:
+                frame = None
+            except Exception:
+                frame = None
+                import traceback
+                logger.error(traceback.format_exc())
+            self._recent_frame = frame
+            if frame:
+                self._frame_size = (frame.width, frame.height)
+                self.last_update = self.g_pool.get_timestamp()
+                events['frame'] = frame
+            elif (self.g_pool.get_timestamp() - self.last_update
+                    > self.ghost_mode_timeout):
+                logger.info('Entering gost mode')
+                if self.online:
+                    self.sensor.unlink()
+                self.sensor = None
+                self._source_id = None
+                self._initial_refresh = True
+                self.update_control_menu()
+                self.last_update = self.g_pool.get_timestamp()
+        else:
+            time.sleep(self.get_frame_timeout/1e3)
 
     def on_notification(self, sensor, event):
+        # should only called if sensor was created
         if self._initial_refresh:
             self.sensor.set_control_value('streaming', True)
             self.sensor.refresh_controls()
@@ -112,39 +154,25 @@ class NDSI_Source(Base_Source):
     def frame_size(self):
         return self._frame_size
 
-    @frame_size.setter
-    def frame_size(self, new_size):
-        self._frame_size = new_size
-
-    def set_frame_size(self, new_size):
-        self.frame_size = new_size
-
     @property
     def frame_rate(self):
-        return 30
-
-    @frame_rate.setter
-    def frame_rate(self, new_rate):
-        pass
+        return self._frame_rate
 
     @property
     def jpeg_support(self):
         return True
 
-    @property
-    def settings(self):
-        settings = super(NDSI_Source, self).settings
-        settings['name'] = self.name
-        settings['sensor_name'] = self.sensor.name
-        settings['host_name'] = self.sensor.host_name
+    def get_init_dict(self):
+        settings = super().get_init_dict()
         settings['frame_rate'] = self.frame_rate
         settings['frame_size'] = self.frame_size
+        if self.online:
+            settings['sensor_name'] = self.sensor.name
+            settings['host_name'] = self.sensor.host_name
+        else:
+            settings['sensor_name'] = self._sensor_name
+            settings['host_name'] = self._host_name
         return settings
-
-    @settings.setter
-    def settings(self, settings):
-        self.frame_size = settings['frame_size']
-        self.frame_rate = settings['frame_rate']
 
     def init_gui(self):
         from pyglui import ui
@@ -220,6 +248,12 @@ class NDSI_Source(Base_Source):
         del self.g_pool.capture_source_menu.elements[:]
         del self.uvc_menu[:]
         self.control_id_ui_mapping = {}
+        if not self.sensor:
+            self.g_pool.capture_source_menu.append(
+                ui.Info_Text(('Sensor %s @ %s not available. '
+                              + 'Running in ghost mode.') % (self._sensor_name,
+                                                             self._host_name)))
+            return
 
         uvc_controls = []
         other_controls = []
@@ -229,7 +263,8 @@ class NDSI_Source(Base_Source):
             else:
                 other_controls.append(entry)
 
-        self.add_controls_to_menu(self.g_pool.capture_source_menu, other_controls)
+        self.add_controls_to_menu(self.g_pool.capture_source_menu,
+                                  other_controls)
         self.add_controls_to_menu(self.uvc_menu, uvc_controls)
         self.g_pool.capture_source_menu.append(self.uvc_menu)
 
@@ -237,9 +272,11 @@ class NDSI_Source(Base_Source):
                                                          self.sensor.reset_all_control_values))
 
     def cleanup(self):
-        self.sensor.unlink()
+        if self.online:
+            self.sensor.unlink()
         self.sensor = None
         self.uvc_menu = None
+        super().cleanup()
 
 
 class NDSI_Manager(Base_Manager):
@@ -253,10 +290,12 @@ class NDSI_Manager(Base_Manager):
     gui_name = 'Pupil Mobile'
 
     def __init__(self, g_pool):
-        super(NDSI_Manager, self).__init__(g_pool)
+        super().__init__(g_pool)
         self.network = ndsi.Network(callbacks=(self.on_event,))
         self.network.start()
         self.selected_host = None
+        self._recover_in = 3
+        self._rejoin_in = 400
 
     def cleanup(self):
         self.deinit_gui()
@@ -271,7 +310,6 @@ class NDSI_Manager(Base_Manager):
             devices = {
                 s['host_uuid']: s['host_name']  # removes duplicates
                 for s in self.network.sensors.values()
-                if s['sensor_type'] == 'video'
             }
             devices = [pair for pair in devices.items()]  # create tuples
             # split tuples into 2 lists
@@ -283,13 +321,11 @@ class NDSI_Manager(Base_Manager):
                 self.re_build_ndsi_menu()
 
         host_sel, host_sel_labels = host_selection_list()
-        ui_elements.append(ui.Selector(
-            'selected_host', self,
-            selection=host_sel,
-            labels=host_sel_labels,
-            setter=view_host,
-            label='Remote host'
-        ))
+        ui_elements.append(ui.Selector('selected_host', self,
+                                       selection=host_sel,
+                                       labels=host_sel_labels,
+                                       setter=view_host,
+                                       label='Remote host'))
 
         self.g_pool.capture_selector_menu.extend(ui_elements)
         if not self.selected_host:
@@ -314,7 +350,6 @@ class NDSI_Manager(Base_Manager):
             if not source_uid:
                 return
             settings = {
-                'source_class_name': NDSI_Source.class_name(),
                 'frame_size': self.g_pool.capture.frame_size,
                 'frame_rate': self.g_pool.capture.frame_rate,
                 'source_id': source_uid
@@ -341,20 +376,27 @@ class NDSI_Manager(Base_Manager):
         while self.network.has_events:
             self.network.handle_event()
 
-    def update(self, frame, events):
+    def recent_events(self, events):
         self.poll_events()
+
+        if (isinstance(self.g_pool.capture, NDSI_Source) and not self.g_pool.capture.sensor):
+            if self._recover_in <= 0:
+                self.recover()
+                self._recover_in = int(2*1e3 / self.g_pool.capture.get_frame_timeout)
+            else:
+                self._recover_in -= 1
+
+            if self._rejoin_in <= 0:
+                logger.debug('Rejoining network...')
+                self.network.rejoin()
+                self._rejoin_in = int(10*1e3 / self.g_pool.capture.get_frame_timeout)
+            else:
+                self._rejoin_in -= 1
 
     def on_event(self, caller, event):
         if event['subject'] == 'detach':
-            logger.debug('detached: {}'.format(event))
-            name = str('{} @ {}'.format(event['sensor_name'], event['host_name']))
-            self.notify_all({
-                'subject': 'capture_manager.source_lost',
-                'source_class_name': NDSI_Source.class_name(),
-                'source_id': event['sensor_uuid'],
-                'name': name
-            })
-            sensors = [s for s in self.network.sensors.values() if s['sensor_type'] == 'video']
+            logger.debug('detached: %s' % event)
+            sensors = [s for s in self.network.sensors.values()]
             if self.selected_host == event['host_uuid']:
                 if sensors:
                     self.selected_host = sensors[0]['host_uuid']
@@ -362,40 +404,23 @@ class NDSI_Manager(Base_Manager):
                     self.selected_host = None
                 self.re_build_ndsi_menu()
 
-        elif (event['subject'] == 'attach'
-              and event['sensor_type'] == 'video'):
-            logger.debug('attached: {}'.format(event))
-            name = '{} @ {}'.format(event['sensor_name'], event['host_name'])
-            self.notify_all({
-                'subject': 'capture_manager.source_found',
-                'source_class_name': NDSI_Source.class_name(),
-                'source_id': event['sensor_uuid'],
-                'name': name
-            })
+        elif event['subject'] == 'attach':
+            if event['sensor_type'] == 'video':
+                logger.debug('attached: {}'.format(event))
+                self.notify_all({'subject': 'capture_manager.source_found'})
             if not self.selected_host:
                 self.selected_host = event['host_uuid']
             self.re_build_ndsi_menu()
 
     def activate_source(self, settings={}):
-        try:
-            capture = NDSI_Source(self.g_pool, network=self.network, **settings)
-        except InitialisationError as init_error:
-            logger.error('NDSI source could not be initialised.')
-            if init_error.message:
-                logger.error(init_error.message)
-            logger.debug('NDSI source init settings:\n\t{}\n\t{}'.format(self.network, settings))
+        settings['network'] = self.network
+        if hasattr(self.g_pool, 'plugins'):
+            self.g_pool.plugins.add(NDSI_Source, args=settings)
         else:
-            self.g_pool.capture.deinit_gui()
-            self.g_pool.capture.cleanup()
-            self.g_pool.capture = None
-            self.g_pool.capture = capture
-            self.g_pool.capture.init_gui()
+            self.g_pool.replace_source(NDSI_Source.__name__, source_settings=settings)
 
     def recover(self):
-        if self.g_pool.capture.class_name() == Fake_Source.class_name():
-            preferred = self.g_pool.capture.preferred_source
-            if preferred['source_class_name'] == NDSI_Source.class_name():
-                self.activate_source(preferred)
+        self.g_pool.capture.recover(self.network)
 
     def on_notify(self, n):
         """Provides UI for the capture selection
@@ -405,7 +430,8 @@ class NDSI_Manager(Base_Manager):
 
         Emmits notifications:
             ``capture_manager.source_found``
-            ``capture_manager.source_lost``
         """
-        if (n['subject'].startswith('capture_manager.source_found')):
+        if (n['subject'].startswith('capture_manager.source_found')
+                and isinstance(self.g_pool.capture, NDSI_Source)
+                and not self.g_pool.capture.sensor):
             self.recover()
